move toast to its own file
