+++ feature/preview/src/main/java/com/google/jetpackcamera/feature/preview/PreviewUiState.kt
@@ -17,6 +17,7 @@ package com.google.jetpackcamera.feature.preview
 
 import android.widget.Toast
 import androidx.camera.core.CameraSelector
+import com.google.jetpackcamera.feature.preview.ui.ToastMessage
 import com.google.jetpackcamera.settings.model.CameraAppSettings
 
 /**
@@ -32,28 +33,6 @@ data class PreviewUiState(
     //todo: remove after implementing post capture screen
     val toastMessageToShow: ToastMessage? = null
 )
-
-/**
- * Helper class containing information used to create a toast.
- *
- * @param message the text to be displayed.
- * @param isLongToast determines if the display time is [Toast.LENGTH_LONG] or [Toast.LENGTH_SHORT]
- * @param testDesc the *unique* description for toast message. i.e. *captureFailedToast* vs *captureSuccessToast*.
- * @property testTag the identifiable resource ID of a Toast on screen.
- */
-class ToastMessage(
-    val message: String,
-    isLongToast: Boolean = false,
-    val testDesc: String = "Toast Message"
-) {
-    val testTag: String = "Toast Message"
-    val toastLength: Int = when (isLongToast) {
-        true -> Toast.LENGTH_LONG
-        false -> Toast.LENGTH_SHORT
-    }
-}
-
-
 
 /**
  * Defines the current state of Video Recording
